Clean-up bokeh server tests
